requirements.txt
requests
pydantic
pytest
fastapi 
uvicorn
openai 
python-dotenv